CI: ALCF GitLab CI configuration
This commit implements a `.gitlab-ci.yml` suitable for use on the
GitLab instance hosted at the ALCF.  This is written to be general
enough to be hosted elsewhere.  This includes the possibility of
running on other GitLab instances *simultaneously* if one adds
additional `if/when` conditionals to ensure exactly one instance
attempts launch of any given job.

This configuration includes:

 + 5 jobs testing different PrgEnv environments on Theta
 + 3 jobs testing GNU, LLVM and Intel compilers on a Theta login node
     [ disabled by default, since Dirac provides better coverage]
 + 8 jobs on a VM configured like a Dirac node, covering the floor
     and latest of GNU, LLVM, Intel and PGI compilers
 + 2 on macos VMs, covering the current and floor Xcode releases
 + 5 on a Linux/PPC64le VM, covering the floor and latest of GNU
     and LLVM compilers, and the most recent PGI Community Edition
 + ====
 + 23 jobs total

Relatively simple additions are necessary in this file to add new
runners (hosts) and jobs to be run on them.