Makefile: improve dependency tracking
This commit introduces a level of indirection between the generation
and inclusion of files used by GNU make to track dependencies, which
resolves issue 370.

The `No rule to make target` errors reported in issue 370 were
generated at the `include $(wildcard *.d)` as it is GNU make's policy
to ensure all includes are up-to-date before inclusion.  By using a
level of indirection, we introduce a "shadow" dependency file for
which we can catch the case of broken dependencies, and simply force
regeneration.  A shadow file is copied to become the included file
each time it is updated.  Since the included files no longer have any
explicit dependencies, they cannot lead to failures at include-time,
even if they name removed files.

Additionally, a shadow dependency file now expresses *only* itself
(not the object or executable) as dependent upon the headers reached
via preprocess.  The object or executable now depends on the
dependency file, and thus only indirectly on its contents.  However,
since the dependency file is rebuilt anytime the files it names
change, this is sufficient.  This change reduces the volume of
dependency information fed to GNU make, eliminates a hack used for
multiple-target dependencies from PGI compilers, and generally
simplifies the compiler-specific dependency generation "bits".

The net result is that moving between two branches which both have
this commit as ancestors will "do the right thing" with respect to
dependency tracking in the case header files have been removed.

For movement from older branches (such as the 2020.3.0 release) to one
containing this work, it is not possible to *entirely* fix the problem
reported as issue 370.  Empirically it has been observed that
non-parallel builds work, but this is more a "happy accident" than
something designed into the solution.

Where previously the generation of the initial dependency files and
object files was concurrent, these are now two distinct steps (with
the first competing before the second can begin).  However, each step
exposes its own full concurrency and no measurable slow down was
observed over NFS on Dirac.  On a slower shared filesystem such as on
Cori, the cost of additional `stat` calls is small, but measurable.

This restructuring of the dependency tracking introduces an
interesting possibility for the future.  Currently the dependency
information is generated at the first compile, and each subsequent
compile ensures it is up-to-date.  However, it is not strictly
necessary to construct dependency files on the first compile, since
the lack of any object files is sufficient to see them all built.
With some additional work, it *may* be possible to build the
dependency files only when the source and object both exist.  This
means that the end-user who compiles exactly once would see a
measurable reduction in built time, while the developer would continue
to have dependency tracking beginning with the second compilation.