Makefile: use -MM for dependency generation
This commit replaces `-M` with `-MM` as the preferred (where
supported) flag to use when invoking the compiler to generate
dependency files.  The difference between these two is that the latter
omits system headers, greatly reducing the number of dependencies
which make must process (as by a `stat` system call) for every
compile.

Currently this is enabled for all compiler families except PGI, which
has broken behavior for `-MM`.  Testing confirms sane behavior for the
current and floor versions of GNU, LLVM, Intel, Cray and Apple/Xcode
compiler families.  Insane behavior for PGI has been reconfirmed.