Move ident watermarks to a dedicated source file
The watermarking file is rebuilt on every rebuild of libupcxx.a,
so it should be minimal to reduce re-compilation time for maintainers.

This role was previously played by runtime.o (our largest
single object file and source file, by a factor of 2x-10x) and has
now been relocated to a new otherwise-empty watermark.o.