Add timeout wrapper to run-tests, when available
This helps ensure that runaway tests don't cause run-tests targets
to hang forever, especially in CI.

The new support uses the `timeout` command (part of GNU coreutils and
commonly available by default on Linux), and is enabled by default
when the timeout command appears to be available and working.

Add exitcode to run-test failure output, and include ERROR lines
in the output file grep.