configure: diagnose non-bootstrapped GEX sources
This commit makes small additions to the logic for `--with-gasnet=` in
order to provide a more useful/instructive error message in the cases
of an unbootstrapped source tree or tarball (such as would occur for a
fresh clone or a tarball built from one).

Previously these cases were either reported as NOT an instance of
GASNet sources (for a source tarball), or were accepted by UPC++
configure (for a source directory) only to fail the subsequent `make`
due to the lack of a GASNet configure script.

Additionally, this corrects a bug in which error messages related to
the `--with-gasnet...` option were sometimes missing the value:
  `GASNET='' does not appear to contain GASNet`
Where the option value _should_ appear within the single quotes.