test/alloc: Change default max allocation count to 2^10
This reduces the runtime for dirac/smp/debug with defaults from over 2
minutes to under 5 seconds, without sacrificing any useful coverage.
The command line can still be used to tweak the parameter.

harness was already explicitly requesting 2^10 for all variants of
this test, but dev-check was using the default and taking longer
that necessary (and timing out when using valgrind).