CI scripts: run through all stages
This commit makes adjustments to the "dev-ci" scripts to ensure that a
failure in a `make dev-tests` or `make `dev-run-tests` does not
prevent running the remaining stages (including the case, such as on
Theta, where debug and opt are split into two run stages).

Failure to configure or build UPC++ remain fatal.