Makefile.tests: Add support for expected compile errors
When source file NAME.cpp has a matching NAME.errc file in the same directory,
that indicates the test is expected to generate a compile error, and will be
considered a failure if compilation succeeds. In addition, the *.errc file
contains a list of perl-compatible regular expressions, one per line,
where at least one pattern must match the output generated by the compiler
in order for the test to succeed. Empty lines and lines starting with
'#' are ignored. To accept any compile error, use the pattern '.'.