issue #392: Prevent silent use of by-value communication APIs for huge types
Add static assertions to enforce a default size limit of 512 bytes
for types passed to by-value APIs for rget, rput, broadcast and reduce.

The correct resolution is for programs to call the bulk overload instead
for statically large objects, but in the interests of backwards
compatibility for legacy codes, the limit can also be raised at
application compile time via `-DUPCXX_MAX_VALUE_SIZE=n`.