Makefile.tests: detect "SKIPPED" compile tests
This commit extends the pass/fail detection logic for test compilation
to include "SKIPPED", as indicated by a non-zero exit from the
compiler and `UPCXX_TEST_SKIPPED` in the output.  These do not count
as failures.

The recommended usage is for known failures, especially cases where
the compiler crashes.  To use, a test file may contain something like:

```
  #if SOME_CONDITION
    #error UPCXX_TEST_SKIPPED [optional explanation].  See issue #XXX.
  #endif
``