fixup: Makefile: eliminate GASNET_VAR(_VAL)
This commit ensures a valid generate makefile fragment is always
included from `gasnet.mak` to correct a regression in which `make
install` would run with an empty `$(CONDUITS)` list in the
`--enable-single=debug` case.

The logic in PR#242 to extract variables such as `$(CONDUIT)` from a
generated fragment (rather that via `$(shell ...)`) had failed to
consider that steps such as `install` do not have settings for
`$(OPTLEV)` and `$(DBGSYM)`.  This commit modifies the logic that sets
`$(GASNET_CODEMODE)` to use `$(UPCXX_DBGOPT)` when it is single-word
(indicating `--enable-single=...`) instead of checking those two.

Note that the bulk of the changed lines in this commit are due to the
addition of indentation to make the code more readable.