detail::launder_unconstructed: Strengthen our compiler assertion
This commit strengthens the implementation for our pre-C++17 replacement
for `std::launder`.

Our previous version was telling the compiler that we may have changed
the input pointer in unspecified ways before returning it, so for
example we might have changed it to point to a validly constructed
object of type `T`. Howver this doesn't achieve our intended effect
if a global analysis can prove that no valid `T` constructor is ever invoked,
or that there are no valid objects of type `T` whose lifetime overlaps
the call to `detail::launder_unconstructed`. Additionally, it inhibits
analysis from inferring the memory address is unchanged, which is unnecessary.

The replacement (based on examples in the
[GCC ASM manual](https://gcc.gnu.org/onlinedocs/gcc/Extended-Asm.html))
instead tells the compiler we are changing the contents of the memory
referenced by the pointer in unspecified ways (which might include
invoking the constructor). This way we are only "killing" information
about the contents of the object, and not about the pointer itself.

The intel compiler ICEs on the simplest generic gnu-style extended asm,
so use a convoluted variant that means the same thing.