GitLab CI: add (manual) runners on Cori
This commit adds the following platforms to `.gitlab-ci.yml`, for
testing on Cori (via manually launched runners):
    cori_hsw.PrgEnv-cray
    cori_hsw.PrgEnv-intel
    cori_hsw.PrgEnv-gnu
    cori_hsw.PrgEnv-cray_floor
    cori_hsw.PrgEnv-intel_floor
    cori_hsw.PrgEnv-gnu_floor

As the names indicate, these are all for Haswell nodes.
Addition of KNL variants is possible with relatively little work,
however the difference in testing latency due to slower runs was
significant in some preliminary trials.

These are NOT run by default since gitlab-runner is NOT running as a
"service" on Cori.  So, these require manual runner launch by Paul.