Uniform make variables for example/cuda_vecadd
This commit updates `example/cuda_vecadd/Makefile` to implement
the set of make variables recently standardized in the `jac3d`
and `cannon` examples of upcxx-extras (PR#24).

Additionally, the `Makefile.summit` is removed.  The new variables
permit easily specifying a GPU architecture, and this Makefile named
for "Summit" was not even passing an architecture appropriate to
the V100 GPUs on Summit.