issue #390: Use -purge-option= to workaround PGI floor debug symbol bugginess
This enables us to maintain debug-mode coverage (without actual debug
symbols) for the PGI floor compiler version for some of the larger
(and thus generally more important) tests in our suite.

This resolves the failures we've recently started getting on the
rpc-ctor-trace test (which has grown), and allows us to re-enable
three previously skipped tests that suffered from this same ICE.

This change is "targetted" in that none of the logic added in this PR
is activated unless PGI_DEBUG_SYMBOLS_BROKEN=1 is set in the make
test environment, and GitLab CI for PGI floor is changed to pass that.
Also, the change is selectively applied only to the tests known to be
affected (but we can easily grow this list as needed).