CI scripts: Support several CI_* pipeline settings
This commit updates the scripts in `utils/dev-ci` to support the
`CI_TESTS`, `CI_NO_TESTS`, `CI_EXTRAFLAGS` and `CI_DEV_CHECK`
variables with the same semantics as in `pipeline.sh`.

The default for `CI_DEV_CHECK` is `1`, which differs from
`pipeline.sh`, but retains the prior behavior of these scripts as well
as matching the behavior of our GitLab CI support.

The new support for `CI_EXTRAFLAGS` and `CI_DEV_CHECK` allows minor
simplifications in `.gitlab-ci.yml`.  As does the observation that
these "dev-ci" scripts already honor `CI_NETWORKS`.