Makefile: repair `make clean tests-clean`
This commit repairs a corner-case regression caused by

    6381f617 Repair `make dev-tests-clean` (issue #406)

Prior to that commit, `make tests-clean` and `make dev-tests-clean`
iterated over all possible networks.  However, that was reduced to
iteration over only the configured networks.  That change had the
unexpected side-effect of "breaking" the commands `make clean
tests-clean` and `make clean dev-tests-clean` such that both leave
test executables behind.  That is due to the `clean` target removing
the generated makefile fragments which set `$(CONDUITS)`.

This commit adds simple logic to fall back to `$(ALL_CONDUITS)` in
the case `$(CONDUITS)` is empty/unset.