Raise minimum g++/libstdc++ on Cray XC to 7.1.0
This commit resolves issue #414 by raising our "floor" version of the
`gcc` environment module on the Cray XC from 6.4.0 to 7.1.0.  Changes
include updates to documentation and to configure-time enforcement.

This commit also contains a tangentially related fix for the behavior
of older `icpc` sometimes reporting an erroneous 0 value for
__GNUC_MINOR__.  We dealt with _part_ of this problem before, but not
the case in which the 0 causes the reported version yielded a full
version number below the floor.