Less specific GitLab tags for Xcode runners.
This commit replaces the literal Xcode versions "8" and "11" with
descriptive "floor" and "latest".  This is in response to the update
from Xcode-11 to Xcode-12 on the Catalina runner, and is intended
to be the last time such an update requires tweaking this file.

(cherry picked from commit 07ddd385c413f932bc98815ec79a9fb8d4694567)