GitLab CI: Resume testing of Theta PrgEnv-cray
This commit updates job `theta.PrgEnv-cray` to utilize a recently
installed non-default `cce/10.0.3` which (unlike the default) is
not broken as reported in GASNet Bug 4153.

The removal of `.rules_manual` from the `extends` array allows runs
of this job to resume.

The addition of a `DEV_CI_COMPILER` variable setting tells the CI
script to load the non-default module.  This can be removed when
the default has advanced to a non-broken version.