GitLab CI / CI scripts: de-version macOS bits
This commit removes references to versions (both numeric and "code
name") of macOS and Xcode.  In addition to edits to comments and
documentation, this include renames of GitLab CI jobs and of "dev-ci"
script files.

This is in preparation for addition of a macOS 11.0 (Big Sur) VM to
our nightly testing pool and a corresponding upgrade to the GitLab
runner VM.  It is intended that these changes will be the *last*
required in association with such upgrades.