Fix dev-check when configured using --without-cuda
This commit corrects the behavior of `make dev-check` and related make
targets by checking `$(UPCXX_CUDA)` for equality with `1` in
`bld/tests.mak`, just as is done in `bld/Makefile`.  This is consistent
with the approach we take with `#if` (not `#ifdef`) in headers such
that `0` and undefined are both "false".

Additionally, the `configure` logic as been updated to ensure that
`UPCXX_CUDA` is `1` or empty, in place of the prior tri-state behavior.

Resolves issue #426