Repair am_send_buffer tiny buffer optimization
When serializing RPC's there is a "tiny buffer" optimization that
attempts to serialize the entire message into a stack-allocated
temporary, thus skipping dynamic buffer allocation overheads for
very small messages that will use the eager protocol.

In releases prior to 2019.9.0 (going back to 2018.3.0), the size of this
tiny buffer was 512 bytes for most RPCs (or up to 1024 bytes for
statically bounded RPC on non-aries conduits).

For 2019.9.0, the serialization backend was overhauled, and
`am_send_buffer` was eventually introduced to manage serialization
buffering with three static cases:

1. RPCs with statically bounded length < 256 bytes
2. Other RPCs with bounded length (most common cases)
3. RPCs with unbounded length (discovered during serialization)

When this split was first introduced (around f460731e), the last two
cases still used a 512-byte tiny buffer to perform the optimization
(the first case uses an exact-size static buffer).

However in dd7b6bf0 (and as released in 2019.9.0) the tiny buffer size
for the last two cases was effectively lowered to 64
(a.k.a. `serialization_align_max`), seemingly due to an accidental logic
error.  I conclude this was accidental because:

1. The unbounded `serialization_writer` immediately consumes 32 bytes of
the tiny buffer for the `hunk_footer`, leaving only 32 bytes for
"payload" in the tiny buffer, which is too small for all but trivial
RPCs (even zero-argument lambdas with no captures serialize to ~24 bytes
of payload), effectively defeating the optimization.

2. The bounded serialization_writer that is not statically under 256 bytes
is unlikely to use a tiny buffer of size 64. There are potential cases of
runtime-bounded RPCs under 64 bytes that could use it, but most
interesting cases in this category are likely to exceed 64 bytes.

This commit restores the tiny buffer to no less than 512 bytes, thus
reactivating the optimization for two important classes of RPCs.

As an important exception, we are careful to avoid overallocating the
tiny buffer for statically bounded payloads.