Statically optimize am_send_buffer for NPAM
Add a specialization of `am_send_buffer` for cases where we statically
know that NPAM is always preferred over FPAM for the eager RPC protocol.
This case is currently active for ibv, aries and smp conduits.

When active, this static knowledge is used to eliminate some branches and
(for RPCs with bounded serialization) also eliminates the static
allocation of stack space for the tiny buffer which is statically known
to be dead.