aries: cap default UPCXX_RPC_EAGER_THRESHOLD to 8kb
aries-conduit MaxMedium defaults to ~4kb, which currently determines the default
UPCXX_RPC_EAGER_THRESHOLD. However aries-conduit can be configured
for larger MaxMedium, and measurements on cori (knl and haswell) show that on
configurations doing that it's generally not profitable to push the
UPCXX_RPC_EAGER_THRESHOLD up beyond about 8kb. Therefore this commit
caps the default UPCXX_RPC_EAGER_THRESHOLD at min(maxmedium, 8kb).

This commit has no effect on default configures of aries-conduit
that do not pass --with-aries-max-medium= to raise the MaxMedium size,
because the UPCXX_RPC_EAGER_THRESHOLD will remain capped by the
default of ~4kb. For --with-aries-max-medium >= 8kb this will set
the default threshold to 8kb, but users can still override the default
using the envvar at runtime.