GitLab CI: renames to get junit/xml reporting
This commit renames all the GitLab CI jobs to remove the use of `.`.
This is a work around for
[this issue](https://gitlab.com/gitlab-org/gitlab/-/issues/254643)
in GitLab, where presence of a `.` in job names breaks rendering of
junit/xml testing reports.

Some additional changes between `-` and `-_` have been made to arrive
at an overall consistent use of `-` as the separator between host and
compiler portions of job names.  For instance, we now use the tokens
`macos_current` and `PrgEnv_gnu`, each of which had previously
contained `-`.