CI testing: add probe for OpenMP support
This commit partially addresses issue 319 by adding an automated probe
for OpenMP support and the corresponding compiler flags.  Support for
OpenMP is an optional feature which only impacts the `dev-tests` and
`dev-check` family of make targets.

The most significant change made in the PR is to automatically set
`UPCXX_HAVE_OPENMP` if the new probe succeeds, thus enabling OpenMP
tests in `dev-check` and related targets.  This change means manually
setting of this variable is no longer *necessary* for well-behaved
systems.  However, setting it explicitly to `1` on a system without
support is sufficient to generate failures on the OpenMP tests since
the required flags will be missing.  Therefore, such settings have
been intentionally preserved in `.gitlab-ci.yml` to prevent any silent
loss of coverage.

Due to issues with both cross-compilation and library search paths at
distributed-application launch, is does not seem worth adding a run
check to this probe.  So, in the case of a platform which can link
OpenMP but cannot run, one may need to begin passing a new (but
undocumented) configure argument `--without-openmp` to prevent new
test failures.  Similarly, one can pass `--with-openmp` to ensure a
build failure if support is not found.  Setting `UPCXX_HAVE_OPENMP` to
an empty string at testing time remains available as a means to
disable the OpenMP tests, and again such usage has been retained in
the various CI scripts.