fix issue #245: persona-example deadlocks when --with-mpsc-queue=biglock
We have an undocumented `--with-mpsc-queue=biglock` option that enables
a low-performance version of our thread-safe LPC queues, which utilizes
a "one big lock" approach where queue access is controlled via mutex.
The idea was to have a version that was so obviously correct that it
could serve as a sanity check against the far more subtle default
lock-free LPC queue implementation that is based on atomic operations.

Unfortunately, there was a serious logic error in the biglock
implementation of `intru_queue<T, intru_queue_safety::mpsc>::burst(max_n, fn)`
that caused it to SILENTLY DISCARD LPCs from the queue whenever it
was called with a queue containing more than `max_n` entries (default 100).
This was causing deadlocks in persona-example whenever sending more than
about 200 LPCs, because some were very likely to get "dropped".

Apply the simplest possible fix which is to ignore the value of `max_n`
and always drain (a snapshot of) the entire queue when called.

Actually implementing a `max_n` limit is possible but could be
prohibitively expensive under the current implementation approach;
given the current queuing operators, the cost of deferment via
re-enqueue would scale linearly with the length of the deferred suffix
of elements (raising the time to process the entire queue to quadratic
in its length). It would also compromise the FIFO fairness of the queue.