fixup: Add some alignment assertions
Remove an incorrect use of a cast in gasnet::rpc_as_lpc::build_eager() that was
attempting to advertise an alignment invariant that is not provided by all
callers. The assertions added in the previous commit revealing this lack of
alignment have also been removed (since they are no longer relevant)

Below is an example from the rput test (built using intel 17.0.2 on
dirac/debug/udp), showing how this invariant is violated by the
`am_long_master_packed_cmd()` caller, who passes a buffer constructed on the
stack with 4-byte alignment.

```
[1] #11 0x00000000006747f5 in upcxx::backend::gasnet::rpc_as_lpc::build_eager<upcxx::backend::gasnet::rpc_as_lpc> (cmd_buf=0x7ffcfd175154, cmd_size=8, cmd_alignment=8) at /home/upcxx/src/backend/gasnet/runtime.cpp:1972
[1] #12 0x0000000000662c90 in (anonymous namespace)::am_long_master_packed_cmd (token=0x1fecbb8, payload=0x7ff096f9d3c0, payload_size=4, reply_cb_lo=32119432, reply_cb_hi=0, cmd_size_align15_level1=524305, a0=-1478484, a1=-1, a2=32765, a3=1, a4=0, a5=484584904, a6=32765, a7=484583936, a8=32765, a9=5458584, a10=0, a11=0, a12=0) at /home/upcxx/src/backend/gasnet/runtime.cpp:2421
[1] #13 0x0000000000921113 in AMUDP_processPacket (buf=0x1fecbb8, isloopback=0) at /home/bxx-intel17.2/bld/GASNet-memory_kinds/other/amudp/amudp_reqrep.cpp:805
...
[1] #22 0x000000000052aadb in main () at /home/upcxx/test/rput.cpp:79
```