Update lpc-ctor-trace for issue #450
Increment the expected move count for the `upcxx::lpc()` of a function
object returning an r-value reference.

Note this does not represent a behavioral regression. It reflects the
addition of the semantically required move of the result data from a
runtime temporary location into the implicit promise backing the future,
a move that was previously skipped as part of returning an incorrect
future type that erroneously contained an rref.