Update Apple compiler detection
Between Xcode versions 10 and 11, Apple changed the version output
from "Apple LLVM" to "Apple clang".  We had not noticed previously
because a later check allows clang >= 4 on any x86_64.

This commit widens the pattern for recognition of "good" Apple
compiler versions to allow for both "LLVM" and "clang".  This is
preferred over the prior behavior both because Apple versions (and
some behaviors) do not align with llvm.org ones, and because this
change moves us closer to support for macOS/aarch64.