Avoid other possible instances of issue #421
Fix another use of the same construct that is suspected to break
the PGI optimizer - value capturing a remote completion and later
passing it (via lvalue reference) to deserialized_value for a loopback
remote completion.

This particular code path was not known to produce an observable
misbeavior in our current set of tests, but it's a less common path
that probably has less coverage and seems in danger of being similarly
affected.