configure: relax assumption on GASNet tarball
This commit modifies logic in `configure` to examine the GASNet
tarball to determine its top-level directory, rather than assuming
it matches the basename of the tarball.

The majority of the diff is swapping of the logic which fetches a URL
or verifies a path, with the logic to check for an existing
destination.  This reordering is required since we can no longer claim
to know the destination prior to fetching the tarball.

This commit addresses Issue 442 "Improve configure behavior for GASNet
archives lacking Bootstrap" such that the reported failure when
given a BitBucket autovivified tarball URL now correctly reports as
follows:
```
$ [...]/configure --with-gasnet=https://bitbucket.org/berkeleylab/gasnet/get/develop.tar.gz
Fetching https://bitbucket.org/berkeleylab/gasnet/get/develop.tar.gz
Unpacking develop.tar.gz

GASNET='https://bitbucket.org/berkeleylab/gasnet/get/develop.tar.gz'
appears to contain an un-bootstraped source tree. If you wish to use
these sources, please unpack the archive, run './Bootstrap' in its
top-level directory, and then re-run this configure passing
'--with-gasnet=...' pointing at the unpacked sources.
```

Work to automate in-place Bootstrap of unpacked sources is NOT
included in this commit.