fixup: Support more delimiters in NETWORKS
This commit fixes a regression introduced in pull request 320 for
the normal case of `make check` on some (many?) systems.  The
issue was the setting of an explicitly empty `$(NETWORKS)` was
preventing is receiving its intended default value.  It is suspected
that a difference in GNU Make versions let this slip by in testing
of the pull request.